+++ submarine-sdk/pysubmarine/setup.py
@@ -29,7 +29,7 @@ setup(
         'sqlalchemy',
         'sqlparse',
         'pymysql',
-        'tensorflow==1.14.0',
+        'tensorflow>=1.14.0,<2.0.0',
     ],
     classifiers=[
         'Intended Audience :: Developers',
